requirements.txt
annotated-types==0.7.0
async-timeout==4.0.3
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
boto3==1.34.144
botocore==1.34.144
Brotli==1.1.0
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
cloudpathlib==0.18.1
confection==0.1.5
crontab==1.0.1
cryptography==42.0.8
cssselect2==0.7.0
cssutils==2.11.1
cymem==2.0.8
decorator==4.4.2
filelock==3.15.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-RQ2==18.3
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
fonttools==4.53.1
freezegun==1.5.1
fsspec==2024.6.1
ftfy==6.2.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-monitoring==2.22.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.64.1
html5lib==1.1
huggingface-hub==0.24.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jwt==1.3.1
langcodes==3.4.0
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
onetimepass==1.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
preshed==3.0.9
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycaption==2.2.11
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.8.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-ffmpeg-video-streaming==0.1.16
python-http-client==3.3.7
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rq==1.16.2
rq-scheduler==0.13.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
segment-analytics-python==2.3.2
sendgrid==6.11.0
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
srt==3.5.3
starkbank-ecdsa==2.2.0
sympy==1.13.0
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.4
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weasyprint==62.3
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
xmltodict==0.13.0
zopfli==0.2.3
